requirements.txt
streamlit
streamlit-chat
requests
langchain
llama_index
pydantic
python-dotenv
Requests
sentence-transformers
